rename sidebarControls to toolbarControls
