+++ src/core/shared_state/index.ts
@@ -8,7 +8,7 @@ export { layersGroupsSettingsAtom } from './layersGroupsSettings';
 export { focusedGeometryAtom } from './focusedGeometry';
 export { mapListenersAtom } from './mapListeners';
 export { currentNotificationAtom } from './currentNotifications';
-export { sideControlsBarAtom } from './sideControlsBar';
+export { toolbarControlsAtom } from './toolbarControls';
 export { currentMapAtom } from './currentMap';
 export { currentApplicationAtom } from './currentApplication';
 export { defaultAppLayersAtom, defaultLayersParamsAtom } from './defaultLayers';
