+++ src/features/create_layer/atoms/layerSideBarButtonController.tsx
@@ -5,8 +5,8 @@ import { i18n } from '~core/localization';
 import {
   controlGroup,
   controlVisualGroup,
-  sideControlsBarAtom,
-} from '~core/shared_state/sideControlsBar';
+  toolbarControlsAtom,
+} from '~core/shared_state/toolbarControls';
 import { userResourceAtom } from '~core/auth/atoms/userResource';
 import { AppFeature } from '~core/auth/types';
 import { editableLayerControllerAtom } from './editableLayerController';
@@ -20,7 +20,7 @@ const sidebarButtonParams = {
   visualGroup: controlVisualGroup.noAnalytics,
   icon: <AddLayer24 />,
   onClick: () => {
-    sideControlsBarAtom.toggleActiveState.dispatch(CREATE_LAYER_CONTROL_ID);
+    toolbarControlsAtom.toggleActiveState.dispatch(CREATE_LAYER_CONTROL_ID);
   },
   onChange: (becomesActive: boolean) => {
     if (becomesActive) {
@@ -49,7 +49,7 @@ export const layerSideBarButtonControllerAtom = createAtom(
       )
         return;
 
-      const currentControls = getUnlistedState(sideControlsBarAtom);
+      const currentControls = getUnlistedState(toolbarControlsAtom);
       // This flag already checked in Main.js, when user logout
       // I need to remove this button, if default user not contain that feature.
       // TODO: Add cleanup hooks for features
@@ -58,7 +58,7 @@ export const layerSideBarButtonControllerAtom = createAtom(
         // But not added
         if (!currentControls[sidebarButtonParams.id]) {
           schedule((dispatch) => {
-            dispatch(sideControlsBarAtom.addControl(sidebarButtonParams));
+            dispatch(toolbarControlsAtom.addControl(sidebarButtonParams));
           });
         }
       } else {
@@ -66,7 +66,7 @@ export const layerSideBarButtonControllerAtom = createAtom(
         // But not removed
         if (currentControls[sidebarButtonParams.id]) {
           schedule((dispatch) => {
-            dispatch(sideControlsBarAtom.removeControl(sidebarButtonParams.id));
+            dispatch(toolbarControlsAtom.removeControl(sidebarButtonParams.id));
           });
         }
       }
