+++ src/features/geometry_uploader/index.tsx
@@ -2,11 +2,11 @@ import { Plus24 } from '@konturio/default-icons';
 import {
   controlGroup,
   controlVisualGroup,
-} from '~core/shared_state/sideControlsBar';
+} from '~core/shared_state/toolbarControls';
 import {
   currentMapAtom,
   currentMapPositionAtom,
-  sideControlsBarAtom,
+  toolbarControlsAtom,
 } from '~core/shared_state';
 import { focusedGeometryAtom } from '~core/shared_state/focusedGeometry';
 import app_config from '~core/app_config';
@@ -20,7 +20,7 @@ import {
 import { askGeoJSONFile } from './askGeoJSONFile';
 
 export function initFileUploader() {
-  sideControlsBarAtom.addControl.dispatch({
+  toolbarControlsAtom.addControl.dispatch({
     id: GEOMETRY_UPLOADER_CONTROL_ID,
     name: GEOMETRY_UPLOADER_CONTROL_NAME,
     title: i18n.t('Focus to uploaded geometry'),
