+++ src/features/map_ruler/index.tsx
@@ -1,5 +1,5 @@
 import { Ruler24 } from '@konturio/default-icons';
-import { sideControlsBarAtom } from '~core/shared_state';
+import { toolbarControlsAtom } from '~core/shared_state';
 import {
   MAP_RULER_CONTROL_ID,
   MAP_RULER_CONTROL_NAME,
@@ -8,7 +8,7 @@ import {
 import {
   controlGroup,
   controlVisualGroup,
-} from '~core/shared_state/sideControlsBar';
+} from '~core/shared_state/toolbarControls';
 import { i18n } from '~core/localization';
 import { createLogicalLayerAtom } from '~core/logical_layers/utils/logicalLayerFabric';
 import { layersRegistryAtom } from '~core/logical_layers/atoms/layersRegistry';
@@ -22,7 +22,7 @@ export function initMapRuler() {
     layersRegistryAtom,
   );
 
-  sideControlsBarAtom.addControl.dispatch({
+  toolbarControlsAtom.addControl.dispatch({
     id: MAP_RULER_CONTROL_ID,
     name: MAP_RULER_CONTROL_NAME,
     title: i18n.t('Ruler'),
@@ -31,7 +31,7 @@ export function initMapRuler() {
     visualGroup: controlVisualGroup.noAnalytics,
     icon: <Ruler24 />,
     onClick: (becomesActive) => {
-      sideControlsBarAtom.toggleActiveState.dispatch(MAP_RULER_CONTROL_ID);
+      toolbarControlsAtom.toggleActiveState.dispatch(MAP_RULER_CONTROL_ID);
     },
     onChange: (becomesActive) => {
       if (becomesActive) {
