+++ src/features/side_bar/components/SideBar/SideBar.tsx
@@ -1,14 +1,14 @@
 import { useAction, useAtom } from '@reatom/react';
 import { ActionsBar, ActionsBarBTN } from '@konturio/ui-kit';
 import { nanoid } from 'nanoid';
-import { sideControlsBarAtom } from '~core/shared_state';
+import { toolbarControlsAtom } from '~core/shared_state';
 import { currentTooltipAtom } from '~core/shared_state/currentTooltip';
 import { controlsOrder } from '../../constants';
 import { sortByPredefinedOrder } from './sortByPredefinedOrder';
 import s from './SideBar.module.css';
 
 export function SideBar() {
-  const [controls] = useAtom(sideControlsBarAtom);
+  const [controls] = useAtom(toolbarControlsAtom);
   const setTooltip = useAction(currentTooltipAtom.setCurrentTooltip);
   const resetTooltip = useAction(currentTooltipAtom.resetCurrentTooltip);
